python/requirements.txt
Flask
mysql-connector-python
pyjwt
python-dotenv
schema
